Day 17: Add part 2 solution (needs robustness fixing)
Used the Z3 SMT solver to solve my exact input for part 2. It's a bad
solution, though, because I only happened to be lucky and get the
smallest solution to the problem on my first run of Z3 on the
command-line after finishing my SMT-LIB file. As soon as I made some
small changes, the model came up with a different answer. Z3 isn't meant
for optimization problems, so I should abandon it here.

Also, refactor and cleanup is probably needed, generally in day 17.