Day 17: Convert part 2 solution to Z3 C# API
The part 2 solution is still not generalized, but the SMT part is more
robust now that it has been converted from an SMT-LIB string literal to
the native C# API for Z3. (Parsing the SMT-LIB string with the C# API
doesn't consider all lines, e.g. `set-logic`, which is why it's better
to use the API.)

I've left the `MkMinimize` statement commented out due to bad
performance. But technically it may be needed if in the future Z3
changes their algorithms and the first result generated is not the
minimum for this problem, anymore.