Day 17: Fix slow SMT performance and fix Z3 warning
Fixes Z3 warning message "WARNING: optimization with quantified
constraints is not supported". The fix is to not declare/define the `f`
function as part of the problem assertions, but rather in C# itself.

This also fixes the performance issue with this problem in Z3. Instead
of running for 16+ seconds, this problem now runs in ~0.5 seconds with
the call to `MkMinimize` uncommented (as it should be).