Day 23: Add solutions
TODO: Part 2 is not generalized. It only happens to work on my input
because a loop limit variable cuts the algorithm off at a time when it
happens to have found my correct solution. But in general this wouldn't
work because the algorithm takes too long to execute without the limit
variable.